Start refactoring triggers processing sysyem
Refactory system using pipeline idea in order manage different actions to do on a QuestEvent container which will contains all needed data from the trigger and from the quest step. For the moment all the refactoring is in the pipeline package, but some classes will be moved to the simplequests-apii project. Finaly, the refactoring concerns alos the managment of the multithreading application in order to make it cleaner , more stable but also flexible